script run at least once on every Thursday.
I need to run some cleanup command once in a week. If it were a server cronjob could do the trick. But, my laptop is usually not up all the time. So, I added this script in my hourly cronjob. and it checks if it is Thursday, then run the job, and switch the flag on. On any other days of the week, it switched to off.